support extended pcre compile options
The number of unused option bits available for use in the
pcre compile functions "options" parameter has almost been
exhausted if the size of an "int" type is only 32 bits.

In fact, there are only two unused bits left when using
an int type that is 32 bits wide.  The two unused bits
are the most significant bits, but using the most
significant bit may be unwise as the type being used
to pass the bits is "int" rather than "unsigned" and
setting the most significant bit would drive the value
negative which could have unforeseen consequences.

However, many of the bits are reserved for internal
use and could potentially be re-used provided they
are then extracted to another variable and cleared.

Therefore define the last remaining unused non-sign bit
as an "extended options are present" bit.

If this bit is set, then any non-public assigned bits
may be reused to pass additional flags provided they
are "C1  Affects compile only" settings that can then
be cleared out and moved to a separate variable.

It is considered an error for only the "extended options
are present" bit to be set without at least one additional
"reused" bit being set.

Since all preexisting versions of the PCRE library complain
and return an error if any non-public API option bits are
set when the compile function is called, attempts by new
code to use the new options with an old shared library will
correctly immediately generate an error return.

While this change does not actually add any new extended
option bits itself, it does add all the necessary checks
and bit manipulations to make them usable and easily added.

Note: This patch is licensed under the same terms as PCRE itself.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>