pcreposix: improve REG_BASIC undefined compatibility
Although the POSIX standard says something like "\T" is
undefined because "T" is not a special character, the
other BRE engines treat it as a plain "T" rather than
leaving it alone (which PCRE_POSIX_BASIC_ESC did).

Change the behavior to match the other engines since
that's allowed by POSIX (it's undefined) to improve
compatibility.

This change ends up making the code a little bit
simpler so it's a win all around.

And since we're in there updating the tests to match,
add another POSIX BRE test to check more '\' escape
behavior.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>