pcreposix: add REG_NOSPEC support
The REG_NOSPEC regcomp() option causes the pattern string
to be treated as a literal.

This is a BSDism, but since it's relatively easy to implement
and it improves regex.h compatibility go ahead and add it.

To implement this the following changes are effected:

1. A new PCRE_VERBATIM option is provided.

   This implicitly prefixes the pattern with \Q and then
   disables recognition of \E thereby causing even \E to
   be taken literally.

2. A new REG_NOSPEC option is provided that maps to PCRE_VERBATIM.

   Although it's "new" to pcreposix, it provides the same
   behavior as the BSD REG_NOSPEC option.

Note: This patch is licensed under the same terms as PCRE itself.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>