pcreposix: make regoff_t more POSIX friendly
The POSIX standard requires that the regoff_t be a signed integer
type that can hold the largest value that can be stored in either a
ptrdiff_t type or a ssize_t type.

We can just use a long except on LLP64 architectures and we need
only check two defines for those in which case we need a long long.

Note: This patch is licensed under the same terms as PCRE itself.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>