pcreposix: make regex_t member re_pcre private
The re_pcre member will no longer always be a "pcre" pointer
when JIT is active.  It MIGHT be a pcre pointer and it might
not.  Change the name to "re_impl" and add a "/* keep out */"
comment.

A magic constant will be used in the future to distinguish so
that broken code which sets it directly will continue to work
at least for a while.

Remove accesses to it from pcretest.c entirely as pcretest.c
has no business inspecting it at all.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>