pcreposix.c: use JIT when available
If the library includes JIT support, attempt to JIT compile the
pattern at regcomp time and if that succeeds then run it using
jit_exec at regexec time.

The JIT stack handling is perhaps not quite as flexible as it could
be, but it should be very flexible and portable this way and still
as fast or faster than just using pcre_exec directly.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>