pcre_jit_compile.c: quietly fall back to pcre_exec
If the regular expression does not have suitable JIT code but
pcre_jit_exec has been called anyway, it's likely the caller really
doesn't understand the API.

Instead of returning the completely unhelpful PCRE_ERROR_JIT_BADOPTION
just fall back to calling pcre_exec instead.

That's what the caller would have done if e was using the API
correctly.  It's also more in line with the API documentation that
claims when JIT can't be used it falls back to non-JIT.

At the same time eliminate a NULL dereference should a NULL value
ever be passed in to pcre_jit_exec.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>