pcreposix.c: always just use pcre_exec
It is possible to allocate a separate JIT stack for use when
executing a JIT pattern.  It's also possible to reuse such a
stack so long as only one pcre_jit_exec is ever using it at
any one time.  This generally means allocating one per thread.

Using the pthread API it's possible to automate this stack
allocation so that the first time pcre_jit_exec is called on
a thread the jit stack is allocated and then is reused until
such time as the thread exits and it's automatically freed via
the pthread_key destructor.

Allocating a new jit stack for each pcre_jit_exec call and then
freeing it immediately afterwards simply will not do as that
results in somewhat horrid performance.

On the other hand, pcre_exec can just be called directly instead
of pcre_jit_exec and it can handle all the stack setup.  It will
use a "local" stack automatically for JIT and perform the automatic
fallback to non-JIT as needed.

It turns out that just calling pcre_exec directly edges out just
by a hair the pthread_once stack allocation technique.

It's also much simpler to code and does not require any specific
knowledge of pthread APIs or support for them either.

Therefore, for the regcomp/regexec implementation just switch
to using pcre_exec only.  JIT will automatically be used on an
automatic local stack all without any hassle or performance penalty.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>